Cleaning up logging APIs, only functional changes was to pad ConsoleSink numbers. #1
