+++ packages/esp-js/src/system/index.ts
@@ -3,5 +3,5 @@ export * from './disposables';
 export * from './models';
 export * from './globalState';
 export * from './guard';
-export * from './logger';
+export * from './logging/logger';
 export * from './utils';
