+++ packages/esp-js/tests/system/loggerTests.ts
@@ -1,4 +1,4 @@
-import {Level, LogEvent, Logger, LoggingConfig} from '../../src/system/logger';
+import {Level, LogEvent, Logger, LoggingConfig} from '../../src/system/logging/logger';
 
 describe('Logger', () => {
     let _logger: Logger,
